requirements.txt
Django==1.4.5
django-tastypie==0.9.12
mimeparse==0.1.3
python-dateutil==2.1
six==1.2.0
wsgiref==0.1.2
